% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compter le nombre de boucles}
\usage{
compter_nombre_boucle(df_velo)
}
\arguments{
\item{df_velo}{Un data.frame de trajets vélos.}
}
\value{
Un entier correspondant au nombre de boucles distinctes.
}
\description{
Compte le nombre de boucles distinctes à partir de la colonne
\code{numero_boucle}.
}
