% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compter le nombre total de trajets}
\usage{
compter_nombre_trajets(df_velo)
}
\arguments{
\item{df_velo}{Un data.frame de trajets vélos.}
}
\value{
Un nombre correspondant au total des trajets.
}
\description{
Calcule la somme de la colonne \code{Total}.
}
