% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-df_velo.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{Trajets vélos à Nantes (Toussaint 2025)}
\format{
Un data.frame avec 6 variables :
\describe{
\item{numero_boucle}{Identifiant de la boucle de comptage}
\item{boucle}{Nom de la boucle}
\item{date}{Date de mesure}
\item{jour_semaine}{Jour de la semaine}
\item{Total}{Nombre de trajets}
\item{probabilite_anomalie}{Indicateur d'anomalie}
}
}
\source{
Nantes Métropole
}
\usage{
df_velo
}
\description{
Ce jeu de données contient les trajets vélos enregistrés à Nantes
pendant les vacances de la Toussaint 2025.
}
\keyword{datasets}
