% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les anomalies des trajets vélos}
\usage{
filtre_anomalie(df_velo)
}
\arguments{
\item{df_velo}{Un data.frame de trajets vélos.}
}
\value{
Un data.frame filtré.
}
\description{
Filtre les observations dont la probabilité de présence d'anomalies est
"Forte" ou "Faible", ainsi que les lignes dont le total est négatif ou
supérieur à 10000.
}
