% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{get_data}
\alias{get_data}
\title{Récupérer les données vélos depuis l'API Nantes}
\usage{
get_data(boucle, mois)
}
\arguments{
\item{boucle}{Numéro de boucle.}

\item{mois}{Mois à filtrer.}
}
\value{
Un data.frame filtré.
}
\description{
Récupérer les données vélos depuis l'API Nantes
}
