% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Représenter la distribution hebdomadaire des trajets}
\usage{
plot_distribution_semaine(df_velo)
}
\arguments{
\item{df_velo}{Un data.frame de trajets vélos.}
}
\value{
Un graphique ggplot.
}
\description{
Produit un graphique en barres du nombre de trajets par jour de la semaine.
}
