% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet maximal}
\usage{
trouver_trajet_max(df_velo)
}
\arguments{
\item{df_velo}{Un data.frame de trajets vélos.}
}
\value{
Un data.frame d'une ligne contenant la boucle, la date, le total
maximal, la moyenne sur la boucle et la moyenne sur la date.
}
\description{
Retourne la ligne correspondant au maximum de trajets, après filtrage des
anomalies, ainsi que deux moyennes de comparaison.
}
