% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets par numéro de boucle}
\usage{
filtrer_trajet(trajet, numero)
}
\arguments{
\item{trajet}{Un data.frame de trajets vélos.}

\item{numero}{Un vecteur de numéros de boucle.}
}
\value{
Un data.frame filtré.
}
\description{
Sélectionne uniquement les lignes correspondant aux numéros de boucle
fournis en entrée.
}
